python-mcp-server/requirements.txt
mcp>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
uvicorn>=0.27.0
starlette>=0.36.0
aiohttp>=3.9.0 